Move integration tests to tests dir
