+++ tests/backup_restore_test.go
@@ -1,4 +1,4 @@
-package run_test
+package tests
 
 import (
 	"io/ioutil"
